add at, scope, leaf
